Manage pgvector secret as part of Helm chart
For automated deployment scenarios, it is preferable for all Kubernetes
resources to be managed as part of a Helm chart rather than imperatively
in the Makefile.

This PR moves the pgvector secret into the pgvector subchart and populates
it with the existing default values. The user is able to set the values
via Helm overrides, or by setting environment variables when deploying
via the Makefile (as is already currently possible).

Signed-off-by: Caleb Xu <caxu@redhat.com>